Re-structure module more logically in prepation for accessors
