+++ geoutils/__init__.py
@@ -5,7 +5,7 @@ GeoUtils is a Python package for the analysis of geospatial data.
 from geoutils import examples, projtools, raster, vector  # noqa
 from geoutils._config import config  # noqa
 from geoutils.raster import Mask, Raster, SatelliteImage  # noqa
-from geoutils.vector import Vector  # noqa
+from geoutils.vector.vector import Vector  # noqa
 
 try:
     from geoutils._version import __version__ as __version__  # noqa
