+++ geoutils/interface/interpolate.py
@@ -1,3 +1,4 @@
+"""Functionalities for interpolating a regular grid at points (raster to point cloud)."""
 from __future__ import annotations
 
 from typing import Any, Callable, Literal, overload
