+++ geoutils/pointcloud/pointcloud.py
@@ -1,0 +1,1 @@
+"""Module for future PointCloud class."""
