+++ geoutils/raster/__init__.py
@@ -1,5 +1,7 @@
 from geoutils.raster.raster import Raster, RasterType, Mask, handled_array_funcs  # noqa isort:skip
 from geoutils.raster.array import *  # noqa
+from geoutils.raster.geotransformations import *  # noqa
+from geoutils.raster.georeferencing import *  # noqa
 from geoutils.raster.multiraster import *  # noqa
 from geoutils.raster.sampling import *  # noqa
 from geoutils.raster.satimg import SatelliteImage  # noqa
