+++ geoutils/raster/georeferencing.py
@@ -1,12 +1,16 @@
+"""
+Functions for manipulating georeferencing of the raster objects.
+"""
 from __future__ import annotations
 
 from typing import Iterable, Literal
+import warnings
 
 import numpy as np
 import rasterio as rio
 
 from geoutils._config import config
-from geoutils._typing import ArrayLike, NDArrayNum
+from geoutils._typing import ArrayLike, NDArrayNum, DTypeLike
 
 
 def _ij2xy(
@@ -170,3 +174,95 @@ def _bounds(transform: rio.transform.Affine, shape: tuple[int, int]) -> rio.coor
     """See description of Raster.bounds."""
 
     return rio.coords.BoundingBox(*rio.transform.array_bounds(height=shape[0], width=shape[1], transform=transform))
+
+
+
+def _cast_pixel_interpretation(
+    area_or_point1: Literal["Area", "Point"] | None, area_or_point2: Literal["Area", "Point"] | None
+) -> Literal["Area", "Point"] | None:
+    """
+    Cast two pixel interpretations and warn if not castable.
+
+    Casts to:
+     - "Area" if both are "Area",
+     - "Point" if both are "Point",
+     - None if any of the interpretation is None, or
+     - None if one is "Area" and the other "Point" (and raises a warning).
+    """
+
+    # If one is None, cast to None
+    if area_or_point1 is None or area_or_point2 is None:
+        area_or_point_out = None
+    # If both are equal and not None
+    elif area_or_point1 == area_or_point2:
+        area_or_point_out = area_or_point1
+    else:
+        area_or_point_out = None
+        msg = (
+            'One raster has a pixel interpretation "Area" and the other "Point". To silence this warning, '
+            "either correct the pixel interpretation of one raster, or deactivate "
+            'warnings of pixel interpretation with geoutils.config["warn_area_or_point"]=False.'
+        )
+        if config["warn_area_or_point"]:
+            warnings.warn(message=msg, category=UserWarning)
+
+    return area_or_point_out
+
+
+# Function to set the default nodata values for any given dtype
+# Similar to GDAL for int types, but without absurdly long nodata values for floats.
+# For unsigned types, the maximum value is chosen (with a max of 99999).
+# For signed types, the minimum value is chosen (with a min of -99999).
+def _default_nodata(dtype: DTypeLike) -> int:
+    """
+    Set the default nodata value for any given dtype, when this is not provided.
+    """
+    default_nodata_lookup = {
+        "uint8": 255,
+        "int8": -128,
+        "uint16": 65535,
+        "int16": -32768,
+        "uint32": 99999,
+        "int32": -99999,
+        "float16": -99999,
+        "float32": -99999,
+        "float64": -99999,
+        "float128": -99999,
+        "longdouble": -99999,  # This is float64 on Windows, float128 on other systems, for compatibility
+    }
+    # Check argument dtype is as expected
+    if not isinstance(dtype, (str, np.dtype, type)):
+        raise TypeError(f"dtype {dtype} not understood.")
+
+    # Convert numpy types to string
+    if isinstance(dtype, type):
+        dtype = np.dtype(dtype).name
+
+    # Convert np.dtype to string
+    if isinstance(dtype, np.dtype):
+        dtype = dtype.name
+
+    if dtype in default_nodata_lookup.keys():
+        return default_nodata_lookup[dtype]
+    else:
+        raise NotImplementedError(f"No default nodata value set for dtype {dtype}.")
+
+
+def _cast_nodata(out_dtype: DTypeLike, nodata: int | float | None) -> int | float | None:
+    """
+    Cast nodata value for output data type to default nodata if incompatible.
+
+    :param out_dtype: Dtype of output array.
+    :param nodata: Nodata value.
+
+    :return: Cast nodata value.
+    """
+
+    if out_dtype == bool:
+        nodata = None
+    if nodata is not None and not rio.dtypes.can_cast_dtype(nodata, out_dtype):
+        nodata = _default_nodata(out_dtype)
+    else:
+        nodata = nodata
+
+    return nodata
