+++ geoutils/vector/__init__.py
@@ -1,0 +1,3 @@
+from geoutils.vector.vector import Vector, VectorType
+from geoutils.vector.geometric import *  # noqa
+from geoutils.vector.geotransformations import *  # noqa
