+++ tests/test_pointcloud/test_pointcloud.py
@@ -1,0 +1,1 @@
+"""Test for future PointCloud class."""
