+++ tests/test_raster/test_raster.py
@@ -22,49 +22,9 @@ from pylint.reporters.text import TextReporter
 import geoutils as gu
 from geoutils import examples
 from geoutils._typing import MArrayNum, NDArrayNum
-from geoutils.misc import resampling_method_from_str
 from geoutils.raster.raster import _default_nodata, _default_rio_attrs
 
 DO_PLOT = False
-
-
-def run_gdal_proximity(
-    input_raster: gu.Raster, target_values: list[float] | None, distunits: str = "GEO"
-) -> NDArrayNum:
-    """Run GDAL's ComputeProximity and return the read numpy array."""
-    # Rasterio strongly recommends against importing gdal along rio, so this is done here instead.
-    from osgeo import gdal, gdalconst
-
-    gdal.UseExceptions()
-
-    # Initiate empty GDAL raster for proximity output
-    drv = gdal.GetDriverByName("MEM")
-    proxy_ds = drv.Create("", input_raster.shape[1], input_raster.shape[0], 1, gdal.GetDataTypeByName("Float32"))
-    proxy_ds.GetRasterBand(1).SetNoDataValue(-9999)
-
-    # Save input in temporary file to read with GDAL
-    # (avoids the nightmare of setting nodata, transform, crs in GDAL format...)
-    with tempfile.TemporaryDirectory() as temp_dir:
-        temp_path = os.path.join(temp_dir, "input.tif")
-        input_raster.save(temp_path)
-        ds_raster_in = gdal.Open(temp_path, gdalconst.GA_ReadOnly)
-
-        # Define GDAL options
-        proximity_options = ["DISTUNITS=" + distunits]
-        if target_values is not None:
-            proximity_options.insert(0, "VALUES=" + ",".join([str(tgt) for tgt in target_values]))
-
-        # Compute proximity
-        gdal.ComputeProximity(ds_raster_in.GetRasterBand(1), proxy_ds.GetRasterBand(1), proximity_options)
-        # Save array
-        proxy_array = proxy_ds.GetRasterBand(1).ReadAsArray().astype("float32")
-        proxy_array[proxy_array == -9999] = np.nan
-
-        # Close GDAL datasets
-        proxy_ds = None
-        ds_raster_in = None
-
-    return proxy_array
 
 
 class TestRaster:
@@ -1048,7 +1008,7 @@ class TestRaster:
 
         # When passing the new array as a NaN ndarray, only the valid data is equal, because masked data is NaN in one
         # case, and -9999 in the other
-        r_arr = gu.raster.get_array_and_mask(r)[0]
+        r_arr = gu.raster.array._get_array_and_mask(r)[0]
         r2 = r.copy(new_array=r_arr)
         assert np.ma.allequal(r.data, r2.data)
         # If a nodata value exists, and we update the NaN pixels to be that nodata value, then the two Rasters should
@@ -1251,609 +1211,6 @@ class TestRaster:
         with pytest.raises(ValueError, match=re.escape(message_raster.format(op_name_assign))):
             rst[mask] = 1
 
-    test_data = [[landsat_b4_path, everest_outlines_path], [aster_dem_path, aster_outlines_path]]
-
-    @pytest.mark.parametrize("data", test_data)  # type: ignore
-    def test_crop(self, data: list[str]) -> None:
-        """Test for crop method, also called by square brackets through __getitem__"""
-
-        raster_path, outlines_path = data
-        r = gu.Raster(raster_path)
-
-        # -- Test with crop_geom being a list/tuple -- ##
-        crop_geom: list[float] = list(r.bounds)
-
-        # Test unloaded inplace cropping conserves the shape
-        r.crop(crop_geom=[crop_geom[0] + r.res[0], crop_geom[1], crop_geom[2], crop_geom[3]], inplace=True)
-        assert len(r.data.shape) == 2
-
-        r = gu.Raster(raster_path)
-
-        # Test with same bounds -> should be the same #
-        crop_geom2 = [crop_geom[0], crop_geom[1], crop_geom[2], crop_geom[3]]
-        r_cropped = r.crop(crop_geom2)
-        assert r_cropped.raster_equal(r)
-
-        # - Test cropping each side by a random integer of pixels - #
-        rng = np.random.default_rng(42)
-        rand_int = rng.integers(1, min(r.shape) - 1)
-
-        # Left
-        crop_geom2 = [crop_geom[0] + rand_int * r.res[0], crop_geom[1], crop_geom[2], crop_geom[3]]
-        r_cropped = r.crop(crop_geom2)
-        assert list(r_cropped.bounds) == crop_geom2
-        assert np.array_equal(r.data[:, rand_int:].data, r_cropped.data.data, equal_nan=True)
-        assert np.array_equal(r.data[:, rand_int:].mask, r_cropped.data.mask)
-
-        # Right
-        crop_geom2 = [crop_geom[0], crop_geom[1], crop_geom[2] - rand_int * r.res[0], crop_geom[3]]
-        r_cropped = r.crop(crop_geom2)
-        assert list(r_cropped.bounds) == crop_geom2
-        assert np.array_equal(r.data[:, :-rand_int].data, r_cropped.data.data, equal_nan=True)
-        assert np.array_equal(r.data[:, :-rand_int].mask, r_cropped.data.mask)
-
-        # Bottom
-        crop_geom2 = [crop_geom[0], crop_geom[1] + rand_int * abs(r.res[1]), crop_geom[2], crop_geom[3]]
-        r_cropped = r.crop(crop_geom2)
-        assert list(r_cropped.bounds) == crop_geom2
-        assert np.array_equal(r.data[:-rand_int, :].data, r_cropped.data.data, equal_nan=True)
-        assert np.array_equal(r.data[:-rand_int, :].mask, r_cropped.data.mask)
-
-        # Top
-        crop_geom2 = [crop_geom[0], crop_geom[1], crop_geom[2], crop_geom[3] - rand_int * abs(r.res[1])]
-        r_cropped = r.crop(crop_geom2)
-        assert list(r_cropped.bounds) == crop_geom2
-        assert np.array_equal(r.data[rand_int:, :].data, r_cropped.data, equal_nan=True)
-        assert np.array_equal(r.data[rand_int:, :].mask, r_cropped.data.mask)
-
-        # Same but tuple
-        crop_geom3: tuple[float, float, float, float] = (
-            crop_geom[0],
-            crop_geom[1],
-            crop_geom[2],
-            crop_geom[3] - rand_int * r.res[0],
-        )
-        r_cropped = r.crop(crop_geom3)
-        assert list(r_cropped.bounds) == list(crop_geom3)
-        assert np.array_equal(r.data[rand_int:, :].data, r_cropped.data.data, equal_nan=True)
-        assert np.array_equal(r.data[rand_int:, :].mask, r_cropped.data.mask)
-
-        # -- Test with crop_geom being a Raster -- #
-        r_cropped2 = r.crop(r_cropped)
-        assert r_cropped2.raster_equal(r_cropped)
-
-        # Check that bound reprojection is done automatically if the CRS differ
-        with warnings.catch_warnings():
-            warnings.filterwarnings("ignore", category=UserWarning, message="For reprojection, nodata must be set.*")
-
-            r_cropped_reproj = r_cropped.reproject(crs=3857)
-            r_cropped3 = r.crop(r_cropped_reproj)
-
-        # Original CRS bounds can be deformed during transformation, but result should be equivalent to this
-        r_cropped4 = r.crop(crop_geom=r_cropped_reproj.get_bounds_projected(out_crs=r.crs))
-        assert r_cropped3.raster_equal(r_cropped4)
-
-        # -- Test with inplace=True -- #
-        r_copy = r.copy()
-        r_copy.crop(r_cropped, inplace=True)
-        assert r_copy.raster_equal(r_cropped)
-
-        # - Test cropping each side with a non integer pixel, mode='match_pixel' - #
-        rand_float = rng.integers(1, min(r.shape) - 1) + 0.25
-
-        # left
-        crop_geom2 = [crop_geom[0] + rand_float * r.res[0], crop_geom[1], crop_geom[2], crop_geom[3]]
-        r_cropped = r.crop(crop_geom2)
-        assert r.shape[1] - (r_cropped.bounds.right - r_cropped.bounds.left) / r.res[0] == int(rand_float)
-        assert np.array_equal(r.data[:, int(rand_float) :].data, r_cropped.data.data, equal_nan=True)
-        assert np.array_equal(r.data[:, int(rand_float) :].mask, r_cropped.data.mask)
-
-        # right
-        crop_geom2 = [crop_geom[0], crop_geom[1], crop_geom[2] - rand_float * r.res[0], crop_geom[3]]
-        r_cropped = r.crop(crop_geom2)
-        assert r.shape[1] - (r_cropped.bounds.right - r_cropped.bounds.left) / r.res[0] == int(rand_float)
-        assert np.array_equal(r.data[:, : -int(rand_float)].data, r_cropped.data.data, equal_nan=True)
-        assert np.array_equal(r.data[:, : -int(rand_float)].mask, r_cropped.data.mask)
-
-        # bottom
-        crop_geom2 = [crop_geom[0], crop_geom[1] + rand_float * abs(r.res[1]), crop_geom[2], crop_geom[3]]
-        r_cropped = r.crop(crop_geom2)
-        assert r.shape[0] - (r_cropped.bounds.top - r_cropped.bounds.bottom) / r.res[1] == int(rand_float)
-        assert np.array_equal(r.data[: -int(rand_float), :].data, r_cropped.data.data, equal_nan=True)
-        assert np.array_equal(r.data[: -int(rand_float), :].mask, r_cropped.data.mask)
-
-        # top
-        crop_geom2 = [crop_geom[0], crop_geom[1], crop_geom[2], crop_geom[3] - rand_float * abs(r.res[1])]
-        r_cropped = r.crop(crop_geom2)
-        assert r.shape[0] - (r_cropped.bounds.top - r_cropped.bounds.bottom) / r.res[1] == int(rand_float)
-        assert np.array_equal(r.data[int(rand_float) :, :].data, r_cropped.data.data, equal_nan=True)
-        assert np.array_equal(r.data[int(rand_float) :, :].mask, r_cropped.data.mask)
-
-        # -- Test with mode='match_extent' -- #
-        # Test all sides at once, with rand_float less than half the smallest extent
-        # The cropped extent should exactly match the requested extent, res will be changed accordingly
-        rand_float = rng.integers(1, min(r.shape) / 2 - 1) + 0.25
-        crop_geom2 = [
-            crop_geom[0] + rand_float * r.res[0],
-            crop_geom[1] + rand_float * abs(r.res[1]),
-            crop_geom[2] - rand_float * r.res[0],
-            crop_geom[3] - rand_float * abs(r.res[1]),
-        ]
-
-        # Filter warning about nodata not set in reprojection (because match_extent triggers reproject)
-        with warnings.catch_warnings():
-            warnings.filterwarnings("ignore", category=UserWarning, message="For reprojection, nodata must be set.*")
-            r_cropped = r.crop(crop_geom2, mode="match_extent")
-
-        assert list(r_cropped.bounds) == crop_geom2
-        # The change in resolution should be less than what would occur with +/- 1 pixel
-        assert np.all(
-            abs(np.array(r.res) - np.array(r_cropped.res)) < np.array(r.res) / np.array(r_cropped.shape)[::-1]
-        )
-
-        # Filter warning about nodata not set in reprojection (because match_extent triggers reproject)
-        with warnings.catch_warnings():
-            warnings.filterwarnings("ignore", category=UserWarning, message="For reprojection, nodata must be set.*")
-            r_cropped2 = r.crop(r_cropped, mode="match_extent")
-        assert r_cropped2.raster_equal(r_cropped)
-
-        # -- Test with crop_geom being a Vector -- #
-        outlines = gu.Vector(outlines_path)
-
-        # First, we reproject manually the outline
-        outlines_reproj = gu.Vector(outlines.ds.to_crs(r.crs))
-        r_cropped = r.crop(outlines_reproj)
-
-        # Calculate intersection of the two bounding boxes and make sure crop has same bounds
-        win_outlines = rio.windows.from_bounds(*outlines_reproj.bounds, transform=r.transform)
-        win_raster = rio.windows.from_bounds(*r.bounds, transform=r.transform)
-        final_window = win_outlines.intersection(win_raster).round_lengths().round_offsets()
-        new_bounds = rio.windows.bounds(final_window, transform=r.transform)
-        assert list(r_cropped.bounds) == list(new_bounds)
-
-        # Second, we check that bound reprojection is done automatically if the CRS differ
-        r_cropped2 = r.crop(outlines)
-        assert list(r_cropped2.bounds) == list(new_bounds)
-
-        # -- Test crop works as expected even if transform has been modified, e.g. through downsampling -- #
-        # Test that with downsampling, cropping to same bounds result in same raster
-        r = gu.Raster(raster_path, downsample=5)
-        r_test = r.crop(r.bounds)
-        assert r_test.raster_equal(r)
-
-        # - Test that cropping yields the same results whether data is loaded or not -
-        # With integer cropping (left)
-        rand_int = rng.integers(1, min(r.shape) - 1)
-        crop_geom2 = [crop_geom[0] + rand_int * r.res[0], crop_geom[1], crop_geom[2], crop_geom[3]]
-        r = gu.Raster(raster_path, downsample=5, load_data=False)
-        assert not r.is_loaded
-        r_crop_unloaded = r.crop(crop_geom2)
-        r.load()
-        r_crop_loaded = r.crop(crop_geom2)
-        # TODO: the following condition should be met once issue #447 is solved
-        # assert r_crop_unloaded.raster_equal(r_crop_loaded)
-        assert r_crop_unloaded.shape == r_crop_loaded.shape
-        assert r_crop_unloaded.transform == r_crop_loaded.transform
-
-        # With a float number of pixels added to the right, mode 'match_pixel'
-        rand_float = rng.integers(1, min(r.shape) - 1) + 0.25
-        crop_geom2 = [crop_geom[0], crop_geom[1], crop_geom[2] + rand_float * r.res[0], crop_geom[3]]
-        r = gu.Raster(raster_path, downsample=5, load_data=False)
-        assert not r.is_loaded
-        r_crop_unloaded = r.crop(crop_geom2, mode="match_pixel")
-        r.load()
-        r_crop_loaded = r.crop(crop_geom2, mode="match_pixel")
-        # TODO: the following condition should be met once issue #447 is solved
-        # assert r_crop_unloaded.raster_equal(r_crop_loaded)
-        assert r_crop_unloaded.shape == r_crop_loaded.shape
-        assert r_crop_unloaded.transform == r_crop_loaded.transform
-
-        # - Check related to pixel interpretation -
-
-        # Check warning for a different area_or_point for the match-reference geometry works
-        r.set_area_or_point("Area", shift_area_or_point=False)
-        r2 = r.copy()
-        r2.set_area_or_point("Point", shift_area_or_point=False)
-
-        with pytest.warns(UserWarning, match='One raster has a pixel interpretation "Area" and the other "Point".*'):
-            r.crop(r2)
-
-        # Check that cropping preserves the interpretation
-        crop_geom = [crop_geom[0] + r.res[0], crop_geom[1], crop_geom[2], crop_geom[3]]
-        r_crop = r.crop(crop_geom)
-        assert r_crop.area_or_point == "Area"
-        r2_crop = r2.crop(crop_geom)
-        assert r2_crop.area_or_point == "Point"
-
-    @pytest.mark.parametrize("example", [landsat_b4_path, aster_dem_path, landsat_rgb_path])  # type: ignore
-    def test_translate(self, example: str) -> None:
-        """Test translation works as intended"""
-
-        r = gu.Raster(example)
-
-        # Get original transform
-        orig_transform = r.transform
-        orig_bounds = r.bounds
-
-        # Shift raster by georeferenced units (default)
-        # Check the default behaviour is not inplace
-        r_notinplace = r.translate(xoff=1, yoff=1)
-        assert isinstance(r_notinplace, gu.Raster)
-
-        # Check inplace
-        r.translate(xoff=1, yoff=1, inplace=True)
-        # Both shifts should have yielded the same transform
-        assert r.transform == r_notinplace.transform
-
-        # Only bounds should change
-        assert orig_transform.c + 1 == r.transform.c
-        assert orig_transform.f + 1 == r.transform.f
-        for attr in ["a", "b", "d", "e"]:
-            assert getattr(orig_transform, attr) == getattr(r.transform, attr)
-
-        assert orig_bounds.left + 1 == r.bounds.left
-        assert orig_bounds.right + 1 == r.bounds.right
-        assert orig_bounds.bottom + 1 == r.bounds.bottom
-        assert orig_bounds.top + 1 == r.bounds.top
-
-        # Shift raster using pixel units
-        orig_transform = r.transform
-        orig_bounds = r.bounds
-        orig_res = r.res
-        r.translate(xoff=1, yoff=1, distance_unit="pixel", inplace=True)
-
-        # Only bounds should change
-        assert orig_transform.c + 1 * orig_res[0] == r.transform.c
-        assert orig_transform.f + 1 * orig_res[1] == r.transform.f
-        for attr in ["a", "b", "d", "e"]:
-            assert getattr(orig_transform, attr) == getattr(r.transform, attr)
-
-        assert orig_bounds.left + 1 * orig_res[0] == r.bounds.left
-        assert orig_bounds.right + 1 * orig_res[0] == r.bounds.right
-        assert orig_bounds.bottom + 1 * orig_res[1] == r.bounds.bottom
-        assert orig_bounds.top + 1 * orig_res[1] == r.bounds.top
-
-        # Check that an error is raised for a wrong distance_unit
-        with pytest.raises(ValueError, match="Argument 'distance_unit' should be either 'pixel' or 'georeferenced'."):
-            r.translate(xoff=1, yoff=1, distance_unit="wrong_value")  # type: ignore
-
-    @pytest.mark.parametrize("example", [landsat_b4_path, aster_dem_path])  # type: ignore
-    def test_reproject(self, example: str) -> None:
-        warnings.simplefilter("error")
-
-        # Reference raster to be used
-        r = gu.Raster(example)
-
-        # -- Check proper errors are raised if nodata are not set -- #
-        r_nodata = r.copy()
-        r_nodata.set_nodata(None)
-
-        # Make sure at least one pixel is masked for test 1
-        rand_indices = gu.raster.subsample_array(r_nodata.data, 10, return_indices=True)
-        r_nodata.data[rand_indices] = np.ma.masked
-        assert np.count_nonzero(r_nodata.data.mask) > 0
-
-        # make sure at least one pixel is set at default nodata for test
-        default_nodata = _default_nodata(r_nodata.dtype)
-        rand_indices = gu.raster.subsample_array(r_nodata.data, 10, return_indices=True)
-        r_nodata.data[rand_indices] = default_nodata
-        assert np.count_nonzero(r_nodata.data == default_nodata) > 0
-
-        # 1 - if no force_source_nodata is set and masked values exist, raises an error
-        with pytest.raises(
-            ValueError,
-            match=re.escape(
-                "No nodata set, set one for the raster with self.set_nodata() or use a "
-                "temporary one with `force_source_nodata`."
-            ),
-        ):
-            _ = r_nodata.reproject(res=r_nodata.res[0] / 2, nodata=0)
-
-        # 2 - if no nodata is set and default value conflicts with existing value, a warning is raised
-        with pytest.warns(
-            UserWarning,
-            match=re.escape(
-                f"For reprojection, nodata must be set. Default chosen value "
-                f"{_default_nodata(r_nodata.dtype)} exists in self.data. This may have unexpected "
-                f"consequences. Consider setting a different nodata with self.set_nodata()."
-            ),
-        ):
-            r_test = r_nodata.reproject(res=r_nodata.res[0] / 2, force_source_nodata=default_nodata)
-        assert r_test.nodata == default_nodata
-
-        # 3 - if default nodata does not conflict, should not raise a warning
-        r_nodata.data[r_nodata.data == default_nodata] = 3
-        r_test = r_nodata.reproject(res=r_nodata.res[0] / 2, force_source_nodata=default_nodata)
-        assert r_test.nodata == default_nodata
-
-        # -- Test setting each combination of georeferences bounds, res and size -- #
-
-        # specific for the landsat test case, default nodata 255 cannot be used (see above), so use 0
-        if r.nodata is None:
-            r.set_nodata(0)
-
-        # - Test size - this should modify the shape, and hence resolution, but not the bounds -
-        out_size = (r.shape[1] // 2, r.shape[0] // 2)  # Outsize is (ncol, nrow)
-        r_test = r.reproject(grid_size=out_size)
-        assert r_test.shape == (out_size[1], out_size[0])
-        assert r_test.res != r.res
-        assert r_test.bounds == r.bounds
-
-        # - Test bounds -
-        # if bounds is a multiple of res, outptut res should be preserved
-        bounds = np.copy(r.bounds)
-        dst_bounds = rio.coords.BoundingBox(
-            left=bounds[0], bottom=bounds[1] + r.res[0], right=bounds[2] - 2 * r.res[1], top=bounds[3]
-        )
-        r_test = r.reproject(bounds=dst_bounds)
-        assert r_test.bounds == dst_bounds
-        assert r_test.res == r.res
-
-        # Create bounds with 1/2 and 1/3 pixel extra on the right/bottom.
-        bounds = np.copy(r.bounds)
-        dst_bounds = rio.coords.BoundingBox(
-            left=bounds[0], bottom=bounds[1] - r.res[0] / 3.0, right=bounds[2] + r.res[1] / 2.0, top=bounds[3]
-        )
-
-        # If bounds are not a multiple of res, the latter will be updated accordingly
-        r_test = r.reproject(bounds=dst_bounds)
-        assert r_test.bounds == dst_bounds
-        assert r_test.res != r.res
-
-        # - Test size and bounds -
-        r_test = r.reproject(grid_size=out_size, bounds=dst_bounds)
-        assert r_test.shape == (out_size[1], out_size[0])
-        assert r_test.bounds == dst_bounds
-
-        # - Test res -
-        # Using a single value, output res will be enforced, resolution will be different
-        res_single = r.res[0] * 2
-        r_test = r.reproject(res=res_single)
-        assert r_test.res == (res_single, res_single)
-        assert r_test.shape != r.shape
-
-        # Using a tuple
-        res_tuple = (r.res[0] * 0.5, r.res[1] * 4)
-        r_test = r.reproject(res=res_tuple)
-        assert r_test.res == res_tuple
-        assert r_test.shape != r.shape
-
-        # - Test res and bounds -
-        # Bounds will be enforced for upper-left pixel, but adjusted by up to one pixel for the lower right bound.
-        # for single res value
-        r_test = r.reproject(bounds=dst_bounds, res=res_single)
-        assert r_test.res == (res_single, res_single)
-        assert r_test.bounds.left == dst_bounds.left
-        assert r_test.bounds.top == dst_bounds.top
-        assert np.abs(r_test.bounds.right - dst_bounds.right) < res_single
-        assert np.abs(r_test.bounds.bottom - dst_bounds.bottom) < res_single
-
-        # For tuple
-        r_test = r.reproject(bounds=dst_bounds, res=res_tuple)
-        assert r_test.res == res_tuple
-        assert r_test.bounds.left == dst_bounds.left
-        assert r_test.bounds.top == dst_bounds.top
-        assert np.abs(r_test.bounds.right - dst_bounds.right) < res_tuple[0]
-        assert np.abs(r_test.bounds.bottom - dst_bounds.bottom) < res_tuple[1]
-
-        # - Test crs -
-        out_crs = rio.crs.CRS.from_epsg(4326)
-        r_test = r.reproject(crs=out_crs)
-        assert r_test.crs.to_epsg() == 4326
-
-        # -- Additional tests --
-        # First, make sure dst_bounds extend beyond current extent to create nodata
-        dst_bounds = rio.coords.BoundingBox(
-            left=bounds[0], bottom=bounds[1] - r.res[0], right=bounds[2] + 2 * r.res[1], top=bounds[3]
-        )
-        r_test = r.reproject(bounds=dst_bounds)
-        assert np.count_nonzero(r_test.data.mask) > 0
-
-        # If nodata falls outside the original image range, check range is preserved (with nearest interpolation)
-        r_float = r.astype("float32")  # type: ignore
-        if (r_float.nodata < np.min(r_float)) or (r_float.nodata > np.max(r_float)):
-            r_test = r_float.reproject(bounds=dst_bounds, resampling="nearest")
-            assert r_test.nodata == r_float.nodata
-            assert np.count_nonzero(r_test.data.data == r_test.nodata) > 0  # Some values should be set to nodata
-            assert np.min(r_test.data) == np.min(r_float.data)  # But min and max should not be affected
-            assert np.max(r_test.data) == np.max(r_float.data)
-
-        # Check that nodata works as expected
-        r_test = r_float.reproject(bounds=dst_bounds, nodata=9999)
-        assert r_test.nodata == 9999
-        assert np.count_nonzero(r_test.data.data == r_test.nodata) > 0
-
-        # Test that reproject works the same whether data is already loaded or not
-        assert r.is_loaded
-        r_test1 = r.reproject(crs=out_crs, nodata=0)
-        r_unload = gu.Raster(example, load_data=False)
-        assert not r_unload.is_loaded
-        r_test2 = r_unload.reproject(crs=out_crs, nodata=0)
-        assert r_test1.raster_equal(r_test2)
-
-        # Test that reproject does not fail with resolution as np.integer or np.float types, single value or tuple
-        astype_funcs = [int, np.int32, float, np.float64]
-        for astype_func in astype_funcs:
-            r.reproject(res=astype_func(20.5), nodata=0)
-        for i in range(len(astype_funcs)):
-            for j in range(len(astype_funcs)):
-                r.reproject(res=(astype_funcs[i](20.5), astype_funcs[j](10.5)), nodata=0)
-
-        # Test that reprojection works for several bands
-        for n in [2, 3, 4]:
-            img1 = gu.Raster.from_array(
-                np.ones((n, 500, 500), dtype="uint8"), transform=rio.transform.from_origin(0, 500, 1, 1), crs=4326
-            )
-
-            img2 = gu.Raster.from_array(
-                np.ones((n, 500, 500), dtype="uint8"), transform=rio.transform.from_origin(50, 500, 1, 1), crs=4326
-            )
-
-            out_img = img2.reproject(img1)
-            assert np.shape(out_img.data) == (n, 500, 500)
-            assert (out_img.count, *out_img.shape) == (n, 500, 500)
-
-        # Test that the rounding of resolution is correct for large decimal numbers
-        # (we take an example that used to fail, see issue #354 and #357)
-        data = np.ones((4759, 2453))
-        transform = rio.transform.Affine(
-            24.12423878332849, 0.0, 238286.29553975424, 0.0, -24.12423878332849, 6995453.456051373
-        )
-        crs = rio.CRS.from_epsg(32633)
-        nodata = -9999.0
-        rst = gu.Raster.from_array(data=data, transform=transform, crs=crs, nodata=nodata)
-
-        rst_reproj = rst.reproject(bounds=rst.bounds, res=(20.0, 20.0))
-        # This used to be 19.999999999999999 due to floating point precision
-        assert rst_reproj.res == (20.0, 20.0)
-
-        # -- Test match reference functionalities --
-
-        # - Create 2 artificial rasters -
-        # for r2b, bounds are cropped to the upper left by an integer number of pixels (i.e. crop)
-        # for r2, resolution is also set to 2/3 the input res
-        min_size = min(r.shape)
-        rng = np.random.default_rng(42)
-        rand_int = rng.integers(min_size / 10, min(r.shape) - min_size / 10)
-        new_transform = rio.transform.from_origin(
-            r.bounds.left + rand_int * r.res[0], r.bounds.top - rand_int * abs(r.res[1]), r.res[0], r.res[1]
-        )
-
-        # data is cropped to the same extent
-        new_data = r.data[rand_int::, rand_int::]
-        r2b = gu.Raster.from_array(data=new_data, transform=new_transform, crs=r.crs, nodata=r.nodata)
-
-        # Create a raster with different resolution
-        dst_res = r.res[0] * 2 / 3
-        r2 = r2b.reproject(res=dst_res)
-        assert r2.res == (dst_res, dst_res)
-
-        # Assert the initial rasters are different
-        assert r.bounds != r2b.bounds
-        assert r.shape != r2b.shape
-        assert r.bounds != r2.bounds
-        assert r.shape != r2.shape
-        assert r.res != r2.res
-
-        # Test reprojecting with ref=r2b (i.e. crop) -> output should have same shape, bounds and data, i.e. be the
-        # same object
-        r3 = r.reproject(r2b)
-        assert r3.bounds == r2b.bounds
-        assert r3.shape == r2b.shape
-        assert r3.bounds == r2b.bounds
-        assert r3.transform == r2b.transform
-        assert np.array_equal(r3.data.data, r2b.data.data, equal_nan=True)
-        assert np.array_equal(r3.data.mask, r2b.data.mask)
-
-        if DO_PLOT:
-            fig1, ax1 = plt.subplots()
-            r.plot(ax=ax1, title="Raster 1")
-
-            fig2, ax2 = plt.subplots()
-            r2b.plot(ax=ax2, title="Raster 2")
-
-            fig3, ax3 = plt.subplots()
-            r3.plot(ax=ax3, title="Raster 1 reprojected to Raster 2")
-
-            plt.show()
-
-        # Test reprojecting with ref=r2 -> output should have same shape, bounds and transform
-        # Data should be slightly different due to difference in input resolution
-        r3 = r.reproject(r2)
-        assert r3.bounds == r2.bounds
-        assert r3.shape == r2.shape
-        assert r3.bounds == r2.bounds
-        assert r3.transform == r2.transform
-        assert not np.array_equal(r3.data.data, r2.data.data, equal_nan=True)
-
-        if DO_PLOT:
-            fig1, ax1 = plt.subplots()
-            r.plot(ax=ax1, title="Raster 1")
-
-            fig2, ax2 = plt.subplots()
-            r2.plot(ax=ax2, title="Raster 2")
-
-            fig3, ax3 = plt.subplots()
-            r3.plot(ax=ax3, title="Raster 1 reprojected to Raster 2")
-
-            plt.show()
-
-        # -- Check that if mask is modified afterwards, it is taken into account during reproject -- #
-        # Create a raster with (additional) random gaps
-        r_gaps = r.copy()
-        nsamples = 200
-        rand_indices = gu.raster.subsample_array(r_gaps.data, nsamples, return_indices=True)
-        r_gaps.data[rand_indices] = np.ma.masked
-        assert np.sum(r_gaps.data.mask) - np.sum(r.data.mask) == nsamples  # sanity check
-
-        # reproject raster, and reproject mask. Check that both have same number of masked pixels
-        # TODO: should test other resampling algo
-        r_gaps_reproj = r_gaps.reproject(res=dst_res, resampling="nearest")
-        mask = gu.Raster.from_array(
-            r_gaps.data.mask.astype("uint8"), crs=r_gaps.crs, transform=r_gaps.transform, nodata=None
-        )
-        mask_reproj = mask.reproject(res=dst_res, nodata=255, resampling="nearest")
-        # Final masked pixels are those originally masked (=1) and the values masked during reproject, e.g. edges
-        tot_masked_true = np.count_nonzero(mask_reproj.data.mask) + np.count_nonzero(mask_reproj.data == 1)
-        assert np.count_nonzero(r_gaps_reproj.data.mask) == tot_masked_true
-
-        # If a nodata is set, make sure it is preserved
-        r_nodata = r.copy()
-
-        r_nodata.set_nodata(0)
-
-        r3 = r_nodata.reproject(r2)
-        assert r_nodata.nodata == r3.nodata
-
-        # -- Check inplace behaviour works -- #
-
-        # Check when transform is updated (via res)
-        r_tmp_res = r.copy()
-        r_res = r_tmp_res.reproject(res=r.res[0] / 2)
-        r_tmp_res.reproject(res=r.res[0] / 2, inplace=True)
-
-        assert r_res.raster_equal(r_tmp_res)
-
-        # Check when CRS is updated
-        r_tmp_crs = r.copy()
-        r_crs = r_tmp_crs.reproject(crs=out_crs)
-        r_tmp_crs.reproject(crs=out_crs, inplace=True)
-
-        assert r_crs.raster_equal(r_tmp_crs)
-
-        # -- Test additional errors raised for argument combinations -- #
-
-        # If both ref and crs are set
-        with pytest.raises(ValueError, match=re.escape("Either of `ref` or `crs` must be set. Not both.")):
-            _ = r.reproject(ref=r2, crs=r.crs)
-
-        # Size and res are mutually exclusive
-        with pytest.raises(ValueError, match=re.escape("size and res both specified. Specify only one.")):
-            _ = r.reproject(grid_size=(10, 10), res=50)
-
-        # If wrong type for `ref`
-        with pytest.raises(
-            TypeError, match=re.escape("Type of ref not understood, must be path to file (str), Raster.")
-        ):
-            _ = r.reproject(ref=3)
-
-        # If input reference is string and file and does not exist
-        with pytest.raises(ValueError, match=re.escape("Reference raster does not exist.")):
-            _ = r.reproject(ref="no_file.tif")
-
-        # -- Check warning for area_or_point works -- #
-        r.set_area_or_point("Area", shift_area_or_point=False)
-        r2 = r.copy()
-        r2.set_area_or_point("Point", shift_area_or_point=False)
-
-        with pytest.warns(UserWarning, match='One raster has a pixel interpretation "Area" and the other "Point".*'):
-            r.reproject(r2)
-
-        # Check that reprojecting preserves interpretation
-        r_reproj = r.reproject(res=r.res[0] * 2)
-        assert r_reproj.area_or_point == "Area"
-        r2_reproj = r2.reproject(res=r2.res[0] * 2)
-        assert r2_reproj.area_or_point == "Point"
 
     @pytest.mark.parametrize("example", [landsat_b4_path, aster_dem_path])  # type: ignore
     def test_intersection(self, example: list[str]) -> None:
@@ -2302,7 +1659,7 @@ class TestRaster:
     # The multi-band example will not have a colorbar, so not used in tests
     @pytest.mark.parametrize("example", [landsat_b4_path, landsat_b4_crop_path, aster_dem_path])  # type: ignore
     @pytest.mark.parametrize("figsize", np.arange(2, 20, 2))  # type: ignore
-    def test_show_cbar(self, example, figsize) -> None:
+    def test_plot_cbar(self, example, figsize) -> None:
         """
         Test cbar matches plot height.
         """
@@ -2330,7 +1687,7 @@ class TestRaster:
         # Assert height is the same
         assert h == pytest.approx(h_cbar)
 
-    def test_show(self) -> None:
+    def test_plot(self) -> None:
         # Read single band raster and RGB raster
         img = gu.Raster(self.landsat_b4_path)
         img_RGB = gu.Raster(self.landsat_rgb_path)
@@ -2587,408 +1944,6 @@ class TestRaster:
             red_c.data.data.squeeze().astype("float32"), img.data.data[0, :, :].astype("float32"), equal_nan=True
         )
 
-    def test_resampling_str(self) -> None:
-        """Test that resampling methods can be given as strings instead of rio enums."""
-        warnings.simplefilter("error")
-        assert resampling_method_from_str("nearest") == rio.enums.Resampling.nearest  # noqa
-        assert resampling_method_from_str("cubic_spline") == rio.enums.Resampling.cubic_spline  # noqa
-
-        # Check that odd strings return the appropriate error.
-        try:
-            resampling_method_from_str("CUBIC_SPLINE")  # noqa
-        except ValueError as exception:
-            if "not a valid rasterio.enums.Resampling method" not in str(exception):
-                raise exception
-
-        img1 = gu.Raster(self.landsat_b4_path)
-        img2 = gu.Raster(self.landsat_b4_crop_path)
-        # Set img2 pixel interpretation as "Point" to match "img1" and avoid any warnings
-        img2.set_area_or_point("Point", shift_area_or_point=False)
-        img1.set_nodata(0)
-        img2.set_nodata(0)
-
-        # Resample the rasters using a new resampling method and see that the string and enum gives the same result.
-        img3a = img1.reproject(img2, resampling="q1")
-        img3b = img1.reproject(img2, resampling=rio.enums.Resampling.q1)
-        assert img3a.raster_equal(img3b)
-
-    @pytest.mark.parametrize("example", [landsat_b4_path, aster_dem_path])  # type: ignore
-    def test_polygonize(self, example: str) -> None:
-        """Test that polygonize doesn't raise errors."""
-
-        img = gu.Raster(example)
-
-        # -- Test 1: basic functioning of polygonize --
-
-        # Get unique value for image and the corresponding area
-        value = np.unique(img)[0]
-        pixel_area = np.count_nonzero(img.data == value) * img.res[0] * img.res[1]
-
-        # Polygonize the raster for this value, and compute the total area
-        polygonized = img.polygonize(target_values=value)
-        polygon_area = polygonized.ds.area.sum()
-
-        # Check that these two areas are approximately equal
-        assert polygon_area == pytest.approx(pixel_area)
-        assert isinstance(polygonized, gu.Vector)
-        assert polygonized.crs == img.crs
-
-        # Check default name of data column, and that defining a custom name works the same
-        assert "id" in polygonized.ds.columns
-        polygonized2 = img.polygonize(target_values=value, data_column_name="myname")
-        assert "myname" in polygonized2.ds.columns
-        assert np.array_equal(polygonized2.ds["myname"].values, polygonized.ds["id"].values)
-
-        # -- Test 2: data types --
-
-        # Check that polygonize works as expected for any input dtype (e.g. float64 being not supported by GeoPandas)
-        for dtype in ["uint8", "int8", "uint16", "int16", "uint32", "int32", "float32", "float64"]:
-            img_dtype = img.copy()
-            with warnings.catch_warnings():
-                warnings.filterwarnings(
-                    "ignore", category=UserWarning, message="dtype conversion will result in a " "loss of information.*"
-                )
-                warnings.filterwarnings(
-                    "ignore",
-                    category=UserWarning,
-                    message="Unmasked values equal to the nodata value found in data array.*",
-                )
-                img_dtype = img_dtype.astype(dtype)
-            value = np.unique(img_dtype)[0]
-            img_dtype.polygonize(target_values=value)
-
-        # And for a boolean object, such as a mask
-        mask = img > value
-        mask.polygonize(target_values=1)
-
-    # Test all options, with both an artificial Raster (that has all target values) and a real Raster
-    @pytest.mark.parametrize("distunits", ["GEO", "PIXEL"])  # type: ignore
-    # 0 and 1,2,3 are especially useful for the artificial Raster, and 112 for the real Raster
-    @pytest.mark.parametrize("target_values", [[1, 2, 3], [0], [112], None])  # type: ignore
-    @pytest.mark.parametrize(
-        "raster",
-        [
-            gu.Raster(landsat_b4_path),
-            gu.Raster.from_array(
-                np.arange(25, dtype="int32").reshape(5, 5), transform=rio.transform.from_origin(0, 5, 1, 1), crs=4326
-            ),
-        ],
-    )  # type: ignore
-    def test_proximity_against_gdal(self, distunits: str, target_values: list[float] | None, raster: gu.Raster) -> None:
-        """Test that proximity matches the results of GDAL for any parameter."""
-
-        # TODO: When adding new rasters for tests, specify warning only for Landsat
-        warnings.filterwarnings("ignore", message="Setting default nodata -99999 to mask non-finite values *")
-
-        # We generate proximity with GDAL and GeoUtils
-        gdal_proximity = run_gdal_proximity(raster, target_values=target_values, distunits=distunits)
-        # We translate distunits GDAL option into its GeoUtils equivalent
-        if distunits == "GEO":
-            distance_unit = "georeferenced"
-        else:
-            distance_unit = "pixel"
-        geoutils_proximity = (
-            raster.proximity(distance_unit=distance_unit, target_values=target_values)
-            .data.data.squeeze()
-            .astype("float32")
-        )
-
-        # The results should be the same in all cases
-        try:
-            # In some cases, the proximity differs slightly (generally <1%) for complex settings
-            # (Landsat Raster with target of 112)
-            # It looks like GDAL might not have the right value,
-            # so this particular case is treated differently in tests
-            if target_values is not None and target_values[0] == 112 and raster.filename is not None:
-                # Get index and number of not almost equal point (tolerance of 10-4)
-                ind_not_almost_equal = np.abs(gdal_proximity - geoutils_proximity) > 1e-04
-                nb_not_almost_equal = np.count_nonzero(ind_not_almost_equal)
-                # Check that this is a minority of points (less than 0.5%)
-                assert nb_not_almost_equal < 0.005 * raster.width * raster.height
-
-                # Replace these exceptions by zero in both
-                gdal_proximity[ind_not_almost_equal] = 0.0
-                geoutils_proximity[ind_not_almost_equal] = 0.0
-                # Check that all the rest is almost equal
-                assert np.allclose(gdal_proximity, geoutils_proximity, atol=1e-04, equal_nan=True)
-
-            # Otherwise, results are exactly equal
-            else:
-                assert np.array_equal(gdal_proximity, geoutils_proximity, equal_nan=True)
-
-        # For debugging
-        except Exception as exception:
-            import matplotlib.pyplot as plt
-
-            # Plotting the xdem and GDAL attributes for comparison (plotting "diff" can also help debug)
-            plt.subplot(121)
-            plt.imshow(gdal_proximity)
-            # plt.imshow(np.abs(gdal_proximity - geoutils_proximity)>0.1)
-            plt.colorbar()
-            plt.subplot(122)
-            plt.imshow(geoutils_proximity)
-            # plt.imshow(raster.data.data == 112)
-            plt.colorbar()
-            plt.show()
-
-            # ind_not_equal = np.abs(gdal_proximity - geoutils_proximity)>0.1
-            # print(gdal_proximity[ind_not_equal])
-            # print(geoutils_proximity[ind_not_equal])
-
-            raise exception
-
-    def test_proximity_parameters(self) -> None:
-        """
-        Test that new (different to GDAL's) proximity parameters run.
-        No need to test the results specifically, as those rely entirely on the previous test with GDAL,
-        and tests in rasterize and shapely.
-        #TODO: Maybe add one test with an artificial vector to check it works as intended
-        """
-
-        # -- Test 1: with self's Raster alone --
-        raster1 = gu.Raster(self.landsat_b4_path)
-        prox1 = raster1.proximity()
-
-        # The raster should have the same extent, resolution and CRS
-        assert raster1.georeferenced_grid_equal(prox1)
-
-        # It should change with target values specified
-        prox2 = raster1.proximity(target_values=[255])
-        assert not np.array_equal(prox1.data, prox2.data)
-
-        # -- Test 2: with a vector provided --
-        vector = gu.Vector(self.everest_outlines_path)
-
-        # With default options (boundary geometry)
-        raster1.proximity(vector=vector)
-
-        # With the base geometry
-        raster1.proximity(vector=vector, geometry_type="geometry")
-
-        # With another geometry option
-        raster1.proximity(vector=vector, geometry_type="centroid")
-
-        # With only inside proximity
-        raster1.proximity(vector=vector, in_or_out="in")
-
-    def test_to_pointcloud(self) -> None:
-        """Test to_pointcloud method."""
-
-        # 1/ Single band synthetic data
-
-        # Create a small raster to test point sampling on
-        img_arr = np.arange(25, dtype="int32").reshape(5, 5)
-        img0 = gu.Raster.from_array(img_arr, transform=rio.transform.from_origin(0, 5, 1, 1), crs=4326)
-
-        # Sample the whole raster (fraction==1)
-        points = img0.to_pointcloud()
-        points_arr = img0.to_pointcloud(as_array=True)
-
-        # Check output types
-        assert isinstance(points, gu.Vector)
-        assert isinstance(points_arr, np.ndarray)
-
-        # Check that both outputs (array or vector) are fully consistent, order matters here
-        assert np.array_equal(points.ds.geometry.x.values, points_arr[:, 0])
-        assert np.array_equal(points.ds.geometry.y.values, points_arr[:, 1])
-        assert np.array_equal(points.ds["b1"].values, points_arr[:, 2])
-
-        # Validate that 25 points were sampled (equating to img1.height * img1.width) with x, y, and band0 values.
-        assert points_arr.shape == (25, 3)
-        assert points.ds.shape == (25, 2)  # One less column here due to geometry storing X and Y
-        # Check that X, Y and Z arrays are equal to raster array input independently of value order
-        x_coords, y_coords = img0.ij2xy(i=np.arange(0, 5), j=np.arange(0, 5))
-        assert np.array_equal(np.sort(np.asarray(points_arr[:, 0])), np.sort(np.tile(x_coords, 5)))
-        assert np.array_equal(np.sort(np.asarray(points_arr[:, 1])), np.sort(np.tile(y_coords, 5)))
-        assert np.array_equal(np.sort(np.asarray(points_arr[:, 2])), np.sort(img_arr.ravel()))
-
-        # Check that subsampling works properly
-        points_arr = img0.to_pointcloud(subsample=0.2, as_array=True)
-        assert points_arr.shape == (5, 3)
-
-        # All values should be between 0 and 25
-        assert all(0 <= points_arr[:, 2]) and all(points_arr[:, 2] < 25)
-
-        # 2/ Multi-band synthetic data
-        img_arr = np.arange(25, dtype="int32").reshape(5, 5)
-        img_3d_arr = np.stack((img_arr, 25 + img_arr, 50 + img_arr), axis=0)
-        img3d = gu.Raster.from_array(img_3d_arr, transform=rio.transform.from_origin(0, 5, 1, 1), crs=4326)
-
-        # Sample the whole raster (fraction==1)
-        points = img3d.to_pointcloud(auxiliary_data_bands=[2, 3])
-        points_arr = img3d.to_pointcloud(as_array=True, auxiliary_data_bands=[2, 3])
-
-        # Check equality between both output types
-        assert np.array_equal(points.ds.geometry.x.values, points_arr[:, 0])
-        assert np.array_equal(points.ds.geometry.y.values, points_arr[:, 1])
-        assert np.array_equal(points.ds["b1"].values, points_arr[:, 2])
-        assert np.array_equal(points.ds["b2"].values, points_arr[:, 3])
-        assert np.array_equal(points.ds["b3"].values, points_arr[:, 4])
-
-        # Check it is the right data
-        assert np.array_equal(np.sort(np.asarray(points_arr[:, 0])), np.sort(np.tile(x_coords, 5)))
-        assert np.array_equal(np.sort(np.asarray(points_arr[:, 1])), np.sort(np.tile(y_coords, 5)))
-        assert np.array_equal(np.sort(np.asarray(points_arr[:, 2])), np.sort(img_3d_arr[0, :, :].ravel()))
-        assert np.array_equal(np.sort(np.asarray(points_arr[:, 3])), np.sort(img_3d_arr[1, :, :].ravel()))
-        assert np.array_equal(np.sort(np.asarray(points_arr[:, 4])), np.sort(img_3d_arr[2, :, :].ravel()))
-
-        # With a subsample
-        points_arr = img3d.to_pointcloud(as_array=True, subsample=10, auxiliary_data_bands=[2, 3])
-        assert points_arr.shape == (10, 5)
-
-        # Check the values are still good
-        assert all(0 <= points_arr[:, 2]) and all(points_arr[:, 2] < 25)
-        assert all(25 <= points_arr[:, 3]) and all(points_arr[:, 3] < 50)
-        assert all(50 <= points_arr[:, 4]) and all(points_arr[:, 4] < 75)
-
-        # 3/ Single-band real raster with nodata values
-        img1 = gu.Raster(self.aster_dem_path)
-
-        # Get a large sample to ensure they should be some NaNs normally
-        points_arr = img1.to_pointcloud(subsample=10000, as_array=True, random_state=42)
-        points = img1.to_pointcloud(subsample=10000, random_state=42)
-
-        # This should not load the image
-        assert not img1.is_loaded
-
-        # The subsampled values should be valid and the right shape
-        assert points_arr.shape == (10000, 3)
-        assert points.ds.shape == (10000, 2)  # One less column here due to geometry storing X and Y
-        assert all(np.isfinite(points_arr[:, 2]))
-
-        # The output should respect the default band naming and the input CRS
-        assert np.array_equal(points.ds.columns, ["b1", "geometry"])
-        assert points.crs == img1.crs
-
-        # Try setting the band name
-        points = img1.to_pointcloud(data_column_name="lol", subsample=10)
-        assert np.array_equal(points.ds.columns, ["lol", "geometry"])
-
-        # Keeping the nodata values
-        points_invalid = img1.to_pointcloud(subsample=10000, random_state=42, skip_nodata=False)
-
-        # The subsampled values should not all be valid and the right shape
-        assert points_invalid.ds.shape == (10000, 2)  # One less column here due to geometry storing X and Y
-        assert any(~np.isfinite(points_invalid["b1"].values))
-
-        # 4/ Multi-band real raster
-        img2 = gu.Raster(self.landsat_rgb_path)
-
-        # By default only loads a single band without loading
-        points_arr = img2.to_pointcloud(subsample=10, as_array=True)
-        points = img2.to_pointcloud(subsample=10)
-
-        assert points_arr.shape == (10, 3)
-        assert points.ds.shape == (10, 2)  # One less column here due to geometry storing X and Y
-        assert not img2.is_loaded
-
-        # Storing auxiliary bands
-        points_arr = img2.to_pointcloud(subsample=10, as_array=True, auxiliary_data_bands=[2, 3])
-        points = img2.to_pointcloud(subsample=10, auxiliary_data_bands=[2, 3])
-        assert points_arr.shape == (10, 5)
-        assert points.ds.shape == (10, 4)  # One less column here due to geometry storing X and Y
-        assert not img2.is_loaded
-        assert np.array_equal(points.ds.columns, ["b1", "b2", "b3", "geometry"])
-
-        # Try setting the column name of a specific band while storing all
-        points = img2.to_pointcloud(subsample=10, data_column_name="yes", data_band=2, auxiliary_data_bands=[1, 3])
-        assert np.array_equal(points.ds.columns, ["yes", "b1", "b3", "geometry"])
-
-        # 5/ Error raising
-        with pytest.raises(ValueError, match="Data column name must be a string.*"):
-            img1.to_pointcloud(data_column_name=1)  # type: ignore
-        with pytest.raises(
-            ValueError,
-            match=re.escape("Data band number must be an integer between 1 and the total number of bands (3)."),
-        ):
-            img2.to_pointcloud(data_band=4)
-        with pytest.raises(
-            ValueError, match="Passing auxiliary column names requires passing auxiliary data band numbers as well."
-        ):
-            img2.to_pointcloud(auxiliary_column_names=["a"])
-        with pytest.raises(
-            ValueError, match="Auxiliary data band number must be an iterable containing only integers."
-        ):
-            img2.to_pointcloud(auxiliary_data_bands=[1, 2.5])  # type: ignore
-            img2.to_pointcloud(auxiliary_data_bands="lol")  # type: ignore
-        with pytest.raises(
-            ValueError,
-            match=re.escape("Auxiliary data band numbers must be between 1 and the total number of bands (3)."),
-        ):
-            img2.to_pointcloud(auxiliary_data_bands=[0])
-            img2.to_pointcloud(auxiliary_data_bands=[4])
-        with pytest.raises(
-            ValueError, match=re.escape("Main data band 1 should not be listed in auxiliary data bands [1, 2].")
-        ):
-            img2.to_pointcloud(auxiliary_data_bands=[1, 2])
-        with pytest.raises(ValueError, match="Auxiliary column names must be an iterable containing only strings."):
-            img2.to_pointcloud(auxiliary_data_bands=[2, 3], auxiliary_column_names=["lol", 1])
-        with pytest.raises(
-            ValueError, match="Length of auxiliary column name and data band numbers should be the same*"
-        ):
-            img2.to_pointcloud(auxiliary_data_bands=[2, 3], auxiliary_column_names=["lol", "lol2", "lol3"])
-
-    def test_from_pointcloud(self) -> None:
-        """Test from_pointcloud method."""
-
-        # 1/ Create a small raster to test point sampling on
-        shape = (5, 5)
-        nodata = 100
-        img_arr = np.arange(np.prod(shape), dtype="int32").reshape(shape)
-        transform = rio.transform.from_origin(0, 5, 1, 1)
-        img1 = gu.Raster.from_array(img_arr, transform=transform, crs=4326, nodata=nodata)
-
-        # Check both inputs work (grid coords or transform+shape) on a subsample
-        pc1 = img1.to_pointcloud(subsample=10)
-        img1_sub = gu.Raster.from_pointcloud_regular(pc1, transform=transform, shape=shape)
-
-        grid_coords1 = img1.coords(grid=False)
-        img1_sub2 = gu.Raster.from_pointcloud_regular(pc1, grid_coords=grid_coords1)
-
-        assert img1_sub.raster_equal(img1_sub2)
-
-        # Check that number of valid values are equal to point cloud size
-        assert np.count_nonzero(~img1_sub.data.mask) == 10
-
-        # With no subsampling, should get the exact same raster back
-        pc1_full = img1.to_pointcloud()
-        img1_full = gu.Raster.from_pointcloud_regular(pc1_full, transform=transform, shape=shape, nodata=nodata)
-        assert img1.raster_equal(img1_full, warn_failure_reason=True)
-
-        # 2/ Single-band real raster with nodata values
-        img2 = gu.Raster(self.aster_dem_path)
-        nodata = img2.nodata
-        transform = img2.transform
-        shape = img2.shape
-
-        # Check both inputs work (grid coords or transform+shape) on a subsample
-        pc2 = img2.to_pointcloud(subsample=10000, random_state=42)
-        img2_sub = gu.Raster.from_pointcloud_regular(pc2, transform=transform, shape=shape, nodata=nodata)
-
-        grid_coords2 = img2.coords(grid=False)
-        img2_sub2 = gu.Raster.from_pointcloud_regular(pc2, grid_coords=grid_coords2, nodata=nodata)
-
-        assert img2_sub.raster_equal(img2_sub2, warn_failure_reason=True)
-
-        # Check that number of valid values are equal to point cloud size
-        assert np.count_nonzero(~img2_sub.data.mask) == 10000
-
-        # With no subsampling, should get the exact same raster back
-        pc2_full = img2.to_pointcloud()
-        img2_full = gu.Raster.from_pointcloud_regular(pc2_full, transform=transform, shape=shape, nodata=nodata)
-        assert img2.raster_equal(img2_full, warn_failure_reason=True, strict_masked=False)
-
-        # 3/ Error raising
-        with pytest.raises(TypeError, match="Input grid coordinates must be 1D arrays.*"):
-            gu.Raster.from_pointcloud_regular(pc1, grid_coords=(1, "lol"))  # type: ignore
-        with pytest.raises(ValueError, match="Grid coordinates must be regular*"):
-            grid_coords1[0][0] += 1
-            gu.Raster.from_pointcloud_regular(pc1, grid_coords=grid_coords1)  # type: ignore
-        with pytest.raises(
-            ValueError, match="Either grid coordinates or both geotransform and shape must be provided."
-        ):
-            gu.Raster.from_pointcloud_regular(pc1)
-
 
 class TestMask:
     # Paths to example data
@@ -3146,149 +2101,6 @@ class TestMask:
         assert isinstance(mask, gu.Mask)
         assert np.array_equal(mask.data.data, rst.data.data >= 1)
         assert np.array_equal(mask.data.mask, rst.data.mask)
-
-    @pytest.mark.parametrize("mask", [mask_landsat_b4, mask_aster_dem, mask_everest])  # type: ignore
-    def test_reproject(self, mask: gu.Mask) -> None:
-        # Test 1: with a classic resampling (bilinear)
-
-        # Reproject mask - resample to 100 x 100 grid
-        mask_orig = mask.copy()
-        mask_reproj = mask.reproject(grid_size=(100, 100), force_source_nodata=2)
-
-        # Check instance is respected
-        assert isinstance(mask_reproj, gu.Mask)
-        # Check the dtype of the original mask was properly reconverted
-        assert mask.data.dtype == bool
-        # Check the original mask was not modified during reprojection
-        assert mask_orig.raster_equal(mask)
-
-        # Check inplace behaviour works
-        mask_tmp = mask.copy()
-        mask_tmp.reproject(grid_size=(100, 100), force_source_nodata=2, inplace=True)
-        assert mask_tmp.raster_equal(mask_reproj)
-
-        # This should be equivalent to converting the array to uint8, reprojecting, converting back
-        mask_uint8 = mask.astype("uint8")
-        mask_uint8_reproj = mask_uint8.reproject(grid_size=(100, 100), force_source_nodata=2)
-        mask_uint8_reproj.data = mask_uint8_reproj.data.astype("bool")
-
-        assert mask_reproj.raster_equal(mask_uint8_reproj)
-
-        # Test 2: should raise a warning when the resampling differs from nearest
-
-        with pytest.warns(
-            UserWarning,
-            match="Reprojecting a mask with a resampling method other than 'nearest', "
-            "the boolean array will be converted to float during interpolation.",
-        ):
-            mask.reproject(res=50, resampling="bilinear", force_source_nodata=2)
-
-    @pytest.mark.parametrize("mask", [mask_landsat_b4, mask_aster_dem, mask_everest])  # type: ignore
-    def test_crop(self, mask: gu.Mask) -> None:
-        # Test with same bounds -> should be the same #
-
-        mask_orig = mask.copy()
-        crop_geom = mask.bounds
-        mask_cropped = mask.crop(crop_geom)
-        assert mask_cropped.raster_equal(mask)
-
-        # Check if instance is respected
-        assert isinstance(mask_cropped, gu.Mask)
-        # Check the dtype of the original mask was properly reconverted
-        assert mask.data.dtype == bool
-        # Check the original mask was not modified during cropping
-        assert mask_orig.raster_equal(mask)
-
-        # Check inplace behaviour works
-        mask_tmp = mask.copy()
-        mask_tmp.crop(crop_geom, inplace=True)
-        assert mask_tmp.raster_equal(mask_cropped)
-
-        # - Test cropping each side by a random integer of pixels - #
-        rng = np.random.default_rng(42)
-        rand_int = rng.integers(1, min(mask.shape) - 1)
-
-        # Left
-        crop_geom2 = [crop_geom[0] + rand_int * mask.res[0], crop_geom[1], crop_geom[2], crop_geom[3]]
-        mask_cropped = mask.crop(crop_geom2)
-        assert list(mask_cropped.bounds) == crop_geom2
-        assert np.array_equal(mask.data[:, rand_int:].data, mask_cropped.data.data, equal_nan=True)
-        assert np.array_equal(mask.data[:, rand_int:].mask, mask_cropped.data.mask)
-
-        # Right
-        crop_geom2 = [crop_geom[0], crop_geom[1], crop_geom[2] - rand_int * mask.res[0], crop_geom[3]]
-        mask_cropped = mask.crop(crop_geom2)
-        assert list(mask_cropped.bounds) == crop_geom2
-        assert np.array_equal(mask.data[:, :-rand_int].data, mask_cropped.data.data, equal_nan=True)
-        assert np.array_equal(mask.data[:, :-rand_int].mask, mask_cropped.data.mask)
-
-        # Bottom
-        crop_geom2 = [crop_geom[0], crop_geom[1] + rand_int * abs(mask.res[1]), crop_geom[2], crop_geom[3]]
-        mask_cropped = mask.crop(crop_geom2)
-        assert list(mask_cropped.bounds) == crop_geom2
-        assert np.array_equal(mask.data[:-rand_int, :].data, mask_cropped.data.data, equal_nan=True)
-        assert np.array_equal(mask.data[:-rand_int, :].mask, mask_cropped.data.mask)
-
-        # Top
-        crop_geom2 = [crop_geom[0], crop_geom[1], crop_geom[2], crop_geom[3] - rand_int * abs(mask.res[1])]
-        mask_cropped = mask.crop(crop_geom2)
-        assert list(mask_cropped.bounds) == crop_geom2
-        assert np.array_equal(mask.data[rand_int:, :].data, mask_cropped.data, equal_nan=True)
-        assert np.array_equal(mask.data[rand_int:, :].mask, mask_cropped.data.mask)
-
-        # Test inplace
-        mask_orig = mask.copy()
-        mask_orig.crop(crop_geom2, inplace=True)
-        assert list(mask_orig.bounds) == crop_geom2
-        assert np.array_equal(mask.data[rand_int:, :].data, mask_orig.data, equal_nan=True)
-        assert np.array_equal(mask.data[rand_int:, :].mask, mask_orig.data.mask)
-
-        # Run with match_extent, check that inplace or not yields the same result
-
-        # TODO: Pretty sketchy with the current functioning of "match_extent",
-        #  should we just remove it from Raster.crop() ?
-
-        # mask_cropped = mask.crop(crop_geom2, inplace=False, mode="match_extent")
-        # mask_orig.crop(crop_geom2, mode="match_extent")
-        # assert mask_cropped.raster_equal(mask_orig)
-
-    @pytest.mark.parametrize("mask", [mask_landsat_b4, mask_aster_dem, mask_everest])  # type: ignore
-    def test_polygonize(self, mask: gu.Mask) -> None:
-
-        mask_orig = mask.copy()
-        # Run default
-        vect = mask.polygonize()
-        # Check the dtype of the original mask was properly reconverted
-        assert mask.data.dtype == bool
-        # Check the original mask was not modified during polygonizing
-        assert mask_orig.raster_equal(mask)
-
-        # Check the output is cast into a vector
-        assert isinstance(vect, gu.Vector)
-
-        # Run with zero as target
-        vect = mask.polygonize(target_values=0)
-        assert isinstance(vect, gu.Vector)
-
-        # Check a warning is raised when using a non-boolean value
-        with pytest.warns(UserWarning, match="In-value converted to 1 for polygonizing boolean mask."):
-            mask.polygonize(target_values=2)
-
-    @pytest.mark.parametrize("mask", [mask_landsat_b4, mask_aster_dem, mask_everest])  # type: ignore
-    def test_proximity(self, mask: gu.Mask) -> None:
-
-        mask_orig = mask.copy()
-        # Run default
-        rast = mask.proximity()
-        # Check the dtype of the original mask was properly reconverted
-        assert mask.data.dtype == bool
-        # Check the original mask was not modified during reprojection
-        assert mask_orig.raster_equal(mask)
-
-        # Check that output is cast back into a raster
-        assert isinstance(rast, gu.Raster)
-        # A mask is a raster, so also need to check this
-        assert not isinstance(rast, gu.Mask)
 
     @pytest.mark.parametrize("mask", [mask_landsat_b4, mask_aster_dem, mask_everest])  # type: ignore
     def test_save(self, mask: gu.Mask) -> None:
